requirements.txt
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
defusedxml==0.7.1
Django==4.2.8
django-filter==23.5
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
djlint==1.34.1
draftjs-exporter==2.1.7
EditorConfig==0.12.3
et-xmlfile==1.1.0
filetype==1.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
idna==3.4
jsbeautifier==1.14.11
json5==0.9.14
l18n==2021.3
openpyxl==3.1.2
pathspec==0.12.1
Pillow==9.5.0
pillow-heif==0.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
tomli==2.0.1
tqdm==4.66.1
urllib3==1.26.15
wagtail==5.2.2
webencodings==0.5.1
Willow==1.6.3
